requirements.txt
certifi==2017.4.17
chardet==3.0.4
Django==1.11
django-filter==1.0.4
djangorestframework==3.6.3
idna==2.5
Markdown==2.6.8
psycopg2==2.7.1
pytz==2017.2
requests==2.18.1
urllib3==1.21.1
whitenoise==3.3.0
